+++ tests/adapter/mongo/test_query.py
@@ -109,7 +109,8 @@ def test_build_clinsig(adapter):
             '$in': all_clinsig
             }
 
-def test_build_clinsig_filter(adapter):
+
+def test_build_clinsig_filter(adapter, real_variant_database):
     case_id = 'cust000'
     clinsig_items = [ 4, 5 ]
     all_clinsig = [] # both numerical and human readable values
@@ -133,7 +134,34 @@ def test_build_clinsig_filter(adapter):
           }
         ]
 
-def test_build_clinsig_always(adapter):
+    assert real_variant_database.variant_collection.find_one()
+
+    # Test that the query works with real data:
+
+    case_obj = real_variant_database.case_collection.find_one()
+    case_id = case_obj['_id']
+
+    # Execute a raw query to collect variants that should pass the filter
+    n_results_raw_query = real_variant_database.variant_collection.find({
+        '$and' : [
+            {'genes.region_annotation' : {'$in' : region_annotation}},
+            {'clnsig.value' : {'$in' : [4, 'Likely pathogenic', 5, 'Pathogenic']}},
+            {'case_id' : case_id},
+            {'category' : 'snv'},
+            {'variant_type' : 'clinical'}
+        ]}).count()
+    assert n_results_raw_query
+
+    adapter = real_variant_database
+
+    # filter real variants using query:
+    n_filtered_variants = adapter.variants(case_id=case_id, nr_of_variants=-1, query=query).count()
+
+    # number of variants returned by raw query and filtered variants should be the same:
+    assert n_filtered_variants == n_results_raw_query
+
+
+def test_build_clinsig_always(adapter, real_variant_database):
     case_id = 'cust000'
     clinsig_confident_always_returned = True
     clinsig_items = [ 4, 5 ]
@@ -179,6 +207,44 @@ def test_build_clinsig_always(adapter):
                 }
           }
         ]
+
+    # Test that the query works with real data
+
+    case_obj = real_variant_database.case_collection.find_one()
+    case_id = case_obj['_id']
+
+    adapter = real_variant_database
+    assert adapter.variants(case_id=case_id, nr_of_variants=-1).count()
+
+    # filter variants using query:
+    filtered_variants = list(adapter.variants(case_id=case_id, nr_of_variants=-1, query=query))
+    assert filtered_variants
+
+    # Make sure that variants are filtered as they should:
+    for var in filtered_variants:
+
+        gnomad_filter = False
+        anno_filter = False
+        clisig_filter = False
+
+        if 'gnomad_frequency' in var:
+            if var['gnomad_frequency'] < freq:
+                gnomad_filter = True
+        else:
+            gnomad_filter = True
+
+        for gene in var['genes']:
+            if gene['region_annotation'] in region_annotation:
+                anno_filter = True
+
+        if 'clnsig' in var:
+            for clnsig in var['clnsig']:
+                if clnsig['value'] in [4, 'Likely pathogenic', 5, 'Pathogenic']:
+                    clisig_filter = True
+
+        # Assert that variant passes gnomad filter + anno_filter or has the required clinsig
+        assert (gnomad_filter and anno_filter) or clisig_filter
+
 
 def test_build_spidex_not_reported(adapter):
     case_id = 'cust000'